My merge

+++ spatial-partition/src/main/java/com/iluwatar/spatialpartition/Rect.java
@@ -1,4 +1,4 @@
-/**
+/*
  * The MIT License
  * Copyright © 2014-2019 Ilkka Seppälä
  *
@@ -20,6 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+
 package com.iluwatar.spatialpartition;
 
 /**
@@ -28,14 +29,14 @@ package com.iluwatar.spatialpartition;
  */
 
 public class Rect {
-  int x; 
-  int y; 
-  int width;
-  int height;
+  double x;
+  double y;
+  double width;
+  double height;
 
   //(x,y) - centre of rectangle
 
-  Rect(int x, int y, int width, int height) {
+  Rect(double x, double y, double width, double height) {
     this.x = x;
     this.y = y;
     this.width = width;
